Large modification to version 1.4 - Improved logging mechanism - Improved confirmation of inserting new objects - New services for data publishing - Modifications of publishing services due to changes in DB model

+++ senslog-sdi4apps/src/test/java/cz/hsrs/track/TrackIgnitionTest.java
@@ -91,25 +91,22 @@ public class TrackIgnitionTest {
 		p6.insertToDb();		
 		insObs(p6, 1);
 		
-						
-		TrackData t1 = u.getTrack(unit_id, p1.internalGetTime_stamp());		
+		TrackData t1 = u.getTrack(unit_id, p1.internalGetTime_stamp());
 		Assert.assertEquals(3, u.getTrackLenght(t1.getGid()));
 		
-		TrackData t2 = u.getTrack(unit_id, p5.internalGetTime_stamp());		
+		TrackData t2 = u.getTrack(unit_id, p5.internalGetTime_stamp());
 		//Assert.assertEquals(t2.getEnd(), p6.internalGetTime_stamp());
 		Assert.assertEquals(3, u.getTrackLenght(t2.getGid()));
-		
-		
-		
 	}
-	private void insObs(UnitPosition p, int status) throws SQLException{
+	
+	private void insObs(UnitPosition p, double status) throws SQLException{
 		Observation o = new Observation(p.internalGetTime_stamp(), status , TrackIgnitionSolver.IGNITION_SENSOR_ID, p.getUnit_id());
 		new TrackIgnitionSolver(o).solve();
 	}
+	
 	@After
 	public void delData() throws Exception{
 		util.deleteUnit(unit_id);
-	
 		//util.deleteCompany(comp_id);
 	}
 }
